+++ src/AuthenticatedClient.php
@@ -16,7 +16,7 @@ class AuthenticatedClient implements ClientInterface
      *
      * @var UserSession
      */
-    protected $user;
+    protected $userSession;
 
     /**
      * The client used to access MPX.
@@ -31,13 +31,13 @@ class AuthenticatedClient implements ClientInterface
      * Note that the authentication is not actually established until
      * acquireToken is called.
      *
-     * @param Client      $client The client used to access MPX.
-     * @param UserSession $user   The user associated with this client.
+     * @param Client      $client      The client used to access MPX.
+     * @param UserSession $userSession The user associated with this client.
      */
-    public function __construct(Client $client, UserSession $user)
+    public function __construct(Client $client, UserSession $userSession)
     {
         $this->client = $client;
-        $this->user = $user;
+        $this->userSession = $userSession;
     }
 
     /**
@@ -93,7 +93,7 @@ class AuthenticatedClient implements ClientInterface
         if (!isset($options['query'])) {
             $options['query'] = [];
         }
-        $token = $this->user->acquireToken(null, $reset);
+        $token = $this->userSession->acquireToken(null, $reset);
         $options['query'] += [
             'token' => $token->getValue(),
         ];