+++ src/Service/AccessManagement/ResolveDomain.php
@@ -3,11 +3,8 @@
 namespace Lullabot\Mpx\Service\AccessManagement;
 
 use Cache\Adapter\Common\CacheItem;
-use Cache\Adapter\PHPArray\ArrayCachePool;
-use Lullabot\Mpx\AuthenticatedClient;
 use Lullabot\Mpx\DataService\IdInterface;
 use Lullabot\Mpx\Normalizer\UriNormalizer;
-use Psr\Cache\CacheItemPoolInterface;
 use Symfony\Component\Serializer\Encoder\JsonEncoder;
 use Symfony\Component\Serializer\Normalizer\ArrayDenormalizer;
 use Symfony\Component\Serializer\Normalizer\ObjectNormalizer;
@@ -18,7 +15,7 @@ use Symfony\Component\Serializer\Serializer;
  *
  * @see https://docs.theplatform.com/help/wsf-resolvedomain-method
  */
-class ResolveDomain
+class ResolveDomain extends ResolveBase
 {
     /**
      * The method endpoint (this is not a true REST service).
@@ -29,36 +26,6 @@ class ResolveDomain
      * The schema version of resolveDomain.
      */
     const SCHEMA_VERSION = '1.0';
-
-    /**
-     * The client used to access mpx.
-     *
-     * @var \Lullabot\Mpx\AuthenticatedClient
-     */
-    private $authenticatedClient;
-
-    /**
-     * The cache used to store resolveDomain responses.
-     *
-     * @var CacheItemPoolInterface
-     */
-    private $cache;
-
-    /**
-     * ResolveDomain constructor.
-     *
-     * @param AuthenticatedClient         $authenticatedClient The client used to access mpx.
-     * @param CacheItemPoolInterface|null $cache               (optional) The cache to store responses in. Defaults to an array cache.
-     */
-    public function __construct(AuthenticatedClient $authenticatedClient, CacheItemPoolInterface $cache = null)
-    {
-        $this->authenticatedClient = $authenticatedClient;
-
-        if (!$cache) {
-            $cache = new ArrayCachePool();
-        }
-        $this->cache = $cache;
-    }
 
     /**
      * Resolve all URLs for an account.
@@ -87,14 +54,20 @@ class ResolveDomain
         $response = $this->authenticatedClient->request('GET', static::RESOLVE_DOMAIN_URL, $options);
 
         $encoders = [new JsonEncoder()];
-        $normalizers = [new UriNormalizer(), new ObjectNormalizer(null, null, null, new \Lullabot\Mpx\DataService\ResolveDomainResponseExtractor()), new ArrayDenormalizer()];
+        $normalizers = [new UriNormalizer(), new ObjectNormalizer(null, null, null, new ResolveDomainResponseExtractor()), new ArrayDenormalizer()];
         $serializer = new Serializer($normalizers, $encoders);
 
-        $response = $serializer->deserialize($response->getBody(), ResolveDomainResponse::class, 'json');
+        $resolved = $serializer->deserialize($response->getBody(), ResolveDomainResponse::class, 'json');
         $item = new CacheItem($key);
-        $item->set($response);
+        $item->set($resolved);
+
+        // thePlatform provides no guidance on how long we can cache this for.
+        // Since many of their examples and other mpx clients hardcode these
+        // values, we assume 30 days and that they will implement redirects or
+        // domain aliases if required.
+        $item->expiresAfter(new \DateInterval('P30D'));
         $this->cache->save($item);
 
-        return $response;
+        return $resolved;
     }
 }