+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 
 - Add a method to yield pages in lists #107
+- Service URLs are now cached when loading objects and listening for
+  notifications. #144
 
 ### Changed
 
@@ -16,6 +18,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fix broken range calculations for the end of a list of ranges #113
 - `\Lullabot\Mpx\Service\IdentityManagement\UserInterface` has been added to
   simplify bridging user configuration from applications. #116
+- `\Lullabot\Mpx\Service\AccessManagement\ResolveAllUrls` now follows what
+  `\Lullabot\Mpx\Service\AccessManagement\ResolveDomain` does and has changed
+  it's API as a result. #144
 - The UnixMicrosecondNormalizer class has been renamed to
   UnixMillisecondNormalizer #115